+++ NEWS.md
@@ -1,11 +1,13 @@
-CopernicusMarine v0.4.2.0001
+CopernicusMarine v0.4.2.0002
 -------------
 
  * Added `vignette("proxy")`
+ * Added `vignette("blosc")`
  * In order to pass CRAN checks:
    * Added safeguards to vignette
    * Improved handling of comparing floating point numbers
      when slicing stars_proxy objects.
+   * Added more rigorous checking of availability of required dependencies
  * Added credentials check to all functions downloading
    from S3 storage.